+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+## Version 6.1.8
+
 ## Version 6.1.7
 
 ## Version 6.1.6